+++ src/main/kotlin/org/balhom/currencyprofilesapi/modules/currencyprofiles/domain/props/UpdateCurrencyProfileProps.kt
@@ -1,14 +1,15 @@
 package org.balhom.currencyprofilesapi.modules.currencyprofiles.domain.props
 
+import java.math.BigDecimal
 import java.time.LocalDate
-import java.util.UUID
+import java.util.*
 
 data class UpdateCurrencyProfileProps(
     val id: UUID,
     val userId: UUID,
     val name: String,
-    val balance: Double,
+    val balance: BigDecimal,
     val initDate: LocalDate,
-    val goalMonthlySaving: Double,
-    val goalYearlySaving: Double,
+    val goalMonthlySaving: BigDecimal,
+    val goalYearlySaving: BigDecimal,
 )
