+++ src/main/kotlin/org/balhom/currencyprofilesapi/modules/currencyprofiles/infrastructure/rest/data/responses/CurrencyProfileResponse.kt
@@ -5,18 +5,19 @@ import org.balhom.currencyprofilesapi.common.config.TimeConfig
 import org.balhom.currencyprofilesapi.common.data.enums.CurrencyCodeEnum
 import org.balhom.currencyprofilesapi.common.data.models.FileReferenceData
 import org.balhom.currencyprofilesapi.modules.currencyprofiles.domain.models.CurrencyProfile
+import java.math.BigDecimal
 import java.time.LocalDate
-import java.util.UUID
+import java.util.*
 
 data class CurrencyProfileResponse(
     val id: UUID,
     val name: String,
     val currencyCode: CurrencyCodeEnum,
-    val balance: Double,
+    val balance: BigDecimal,
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = TimeConfig.DATE_ISO_FORMAT)
     val initDate: LocalDate,
-    val goalMonthlySaving: Double,
-    val goalYearlySaving: Double,
+    val goalMonthlySaving: BigDecimal,
+    val goalYearlySaving: BigDecimal,
     val imageData: FileReferenceData?,
     val ownerId: UUID,
 ) {
